Add logging with log4j2
Breaks native image compilation at the very least.
Bumped the log4j2 version to the 2.25.1 which is the latest as of today.
The log4j2 docs say that 2.25.0 onwards has full support for Native Image
compilation, but I was seeing an error related to the initialization of
`org.apache.logging.log4j.util.LazyUtil$SafeLazy`